server/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
pydantic==2.4.2
python-dotenv==1.0.0
flask==2.0.1
flask-cors==3.0.10
werkzeug==2.0.3
gunicorn==20.1.0
motor==3.3.1
pymongo==4.5.0 